removed password.py file
